RavenDB-26838 Extract shared GetSnmpOidsCommand/SnmpEntry test helpers to Tests.Infrastructure
